Rename utils to generator folder
