+++ src/extension/src/utils/generation/generators/CosmosDBGenerator.ts
