+++ src/extension/src/utils/generation/generators/ResourceGroupGenerator.ts
