+++ src/extension/src/utils/generation/generators/TemplatesGenerator.ts
